Pair PR feedback * Removed safelist (until we're confident about the use case with more atomic components) * Moved "global" types into Box and Spacing components, where appropriate * Added support for data attributes in Box
