+++ packages/component-library/components/Spacing/index.ts
@@ -1,2 +1,3 @@
 export * from "./padding"
 export * from "./margin"
+export * from "./types"
