+++ packages/component-library/components/index.ts
@@ -7,3 +7,6 @@ export * from "./MenuList"
 export * from "./Notification"
 export * from "./OffCanvas"
 export * from "./Text"
+export * from "./Spacing"
+export * from "./Heading"
+export * from "./Box"
